gpuav: Only use core Instruction for logging
