+++ layers/gpu/instrumentation/gpuav_shader_instrumentor.h
@@ -18,10 +18,16 @@
 
 #include "containers/custom_containers.h"
 #include "generated/chassis.h"
+#include "state_tracker/shader_instruction.h"
 #include "state_tracker/state_tracker.h"
-#include "gpu/spirv/instruction.h"
 
 #include <vector>
+
+// There is a spirv::Instruction used for normal validation.
+// There is a gpuav::spirv::Instruction that is ONLY intended for shader instrumentation (designed so we can build the shader
+// instrumentation as a seperate library). For logging GPU-AV will want to make use of the normal validaiton instruction class, just
+// alias it with "Instruction" as that name shouldn't collide with anything.
+using Instruction = ::spirv::Instruction;
 
 namespace chassis {
 struct ShaderInstrumentationMetadata;
@@ -165,7 +171,7 @@ class GpuShaderInstrumentor : public ValidationStateTracker {
 
     bool IsSelectiveInstrumentationEnabled(const void *pNext);
 
-    std::string GenerateDebugInfoMessage(VkCommandBuffer commandBuffer, const std::vector<spirv::Instruction> &instructions,
+    std::string GenerateDebugInfoMessage(VkCommandBuffer commandBuffer, const std::vector<Instruction> &instructions,
                                          uint32_t stage_id, uint32_t stage_info_0, uint32_t stage_info_1, uint32_t stage_info_2,
                                          uint32_t instruction_position, const InstrumentedShader *instrumented_shader,
                                          uint32_t shader_id, VkPipelineBindPoint pipeline_bind_point,
