+++ layers/state_tracker/shader_instruction.h
@@ -104,4 +104,6 @@ class Instruction {
 #endif
 };
 
+void GenerateInstructions(const vvl::span<const uint32_t>& spirv, std::vector<Instruction>& instructions);
+
 }  // namespace spirv
